requirements.txt
python-dotenv
soundata
numpy
librosa
soundfile
audioread
h5py
tensorflow
tensorflow-model-optimization
tf_keras